portfolio is now a model with full functionality

+++ migrations/20220908145559-create-portfolio.js
@@ -17,7 +17,7 @@ module.exports = {
         }
       },
       profit: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
 
       createdAt: {
